Development 0.5

+++ model/product_model.go
@@ -1,11 +1,11 @@
 package model
 
 type Product struct {
-	ProductId		string `json:"productId"`
-	CategoryProductId 	string `json:"categoryProductId"`
-	Name              	string `json:"product"`
-	Unit              	string `json:"unit"`
-	Price             	int64  `json:"price"`
-	Stock             	int    `json:"stock"`
-	Audit             	Audit  `json:"audit"`
+	ProductId         string  `json:"productId"`
+	CategoryProductId string  `json:"categoryProductId"`
+	Name              string  `json:"product"`
+	Unit              string  `json:"unit"`
+	Price             float64 `json:"price"`
+	Stock             int     `json:"stock"`
+	Audit             Audit   `json:"audit"`
 }
